Commit Label and Instance factorised in Uri
