+++ phis2-ws/src/main/java/phis2ws/service/resources/ProjectResourceService.java
@@ -48,7 +48,6 @@ import phis2ws.service.view.brapi.form.AbstractResultForm;
 import phis2ws.service.view.brapi.form.ResponseFormGET;
 import phis2ws.service.view.brapi.form.ResponseFormPOST;
 import phis2ws.service.view.brapi.form.ResponseFormProject;
-import phis2ws.service.view.model.phis.Instance;
 import phis2ws.service.view.model.phis.Project;
 
 @Api("/projects")
