+++ phis2-ws/src/main/java/phis2ws/service/view/model/phis/Uri.java
@@ -1,13 +1,13 @@
 //**********************************************************************************************
-//                                       Concept.java 
+//                                       Uri.java 
 //
 // Author(s): Eloan LAGIER
 // PHIS-SILEX version 1.0
-// Copyright © - INRA - 2017
-// Creation date: Decembre, 8 2017
-// Contact: eloan.lagier@inra.fr morgane.vidal@inra.fr, anne.tireau@inra.fr, pascal.neveu@inra.fr
-// Last modification date: January, 25 2018
-// Subject: A class of concept for the service call
+// Copyright © - INRA - 2018
+// Creation date: 26 Feb 2018
+// Contact: eloan.lagier@inra.fr, morgane.vidal@inra.fr, anne.tireau@inra.fr, pascal.neveu@inra.fr
+// Last modification date:  26 Feb, 2018
+// Subject: Uri model for the service call
 //***********************************************************************************************
 package phis2ws.service.view.model.phis;
 
@@ -16,24 +16,15 @@ import java.util.HashMap;
 import java.util.Map;
 
 /**
- * Model of the Concept Object
- *
- * @author lagier
+ * Model of the Uri object
+ * @author Eloan LAGIER
  */
-public class Concept {
-
+public class Uri {
     String uri;
-    //todo
+    String rdfType;
     private Map<String, String> properties = new HashMap<>();
-
-    public Concept() {
-    }
-
-    public Concept(String conceptUri, Map<String, String> infos) {
-        this.uri = conceptUri;
-        this.properties = infos;
-    }
-
+    private Map<String,String> annotations = new HashMap<>();
+    
     @ApiModelProperty(example = "http://www.phenome-fppn.fr/vocabulary/2017#Document")
     public String getUri() {
         return uri;
@@ -42,7 +33,16 @@ public class Concept {
     public void setUri(String uri) {
         this.uri = uri;
     }
+    
+    @ApiModelProperty(example = "Document")
+    public String getRdfType() {
+        return rdfType;
+    }
 
+    public void setRdfType(String rdfType) {
+        this.rdfType = rdfType;
+    }
+    
     @ApiModelProperty(example = "{ Map<String, String> }")
     public Map<String, String> getProperties() {
         return properties;
@@ -50,5 +50,14 @@ public class Concept {
 
     public void addProperty(String key, String value) {
         properties.put(key, value);
+    }
+    
+   @ApiModelProperty(example = "{ Map<String, String> }")
+    public Map<String, String> getAnnotations() {
+        return annotations;
+    }
+
+    public void addAnnotation(String key, String value) {
+        annotations.put(key, value);
     }
 }
